delete review

+++ frontend/src/components/UserListingPage/index.js
@@ -32,8 +32,10 @@ function UserListingPage() {
     if (spots.length == 0) return (<h2>No listings yet.</h2>);
 
     const handleDelete = async (spotId) => {
-        await dispatch(spotsActions.deleteSpot(spotId))
-        // .then(() => history.push('/spots/current'))
+        if (window.confirm('Do you want to delete this spot?')) {
+            await dispatch(spotsActions.deleteSpot(spotId))
+            // .then(() => history.push('/spots/current'))
+        }
     }
 
     return (
